requirements.txt
flask
pymongo
cryptography
pyjwt
werkzeug
flask-mail
web3
python-dotenv
